Move creatable user types to /models
