+++ src/functions/dashboards/get-dashboards-by-group.spec.ts
@@ -7,13 +7,13 @@
  **************************************************************************/
 
 import { random } from 'lodash';
-import { CreatableDashboard, CreatableGroup, isDashboard, User } from '../../models';
+import { CreatableUser, CreatableDashboard, CreatableGroup, isDashboard, User } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { makeLoginOneUser } from '../auth/login-one-user';
 import { makeAddOneUserToManyGroups } from '../groups/add-one-user-to-many-groups';
 import { makeCreateOneGroup } from '../groups/create-one-group';
-import { CreatableUser, makeCreateOneUser, makeGetMyUser, makeGetOneUser } from '../users';
+import { makeCreateOneUser, makeGetMyUser, makeGetOneUser } from '../users';
 import { makeCreateOneDashboard } from './create-one-dashboard';
 import { makeDeleteOneDashboard } from './delete-one-dashboard';
 import { makeGetAllDashboards } from './get-all-dashboards';
