+++ src/functions/groups/remove-one-user-from-one-group.spec.ts
@@ -6,11 +6,11 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { CreatableGroup, User } from '../../models';
+import { CreatableGroup, CreatableUser, User } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { NumericID } from '../../value-objects';
-import { CreatableUser, makeCreateOneUser, makeDeleteOneUser, makeGetOneUser } from '../users';
+import { makeCreateOneUser, makeDeleteOneUser, makeGetOneUser } from '../users';
 import { makeAddOneUserToManyGroups } from './add-one-user-to-many-groups';
 import { makeCreateOneGroup } from './create-one-group';
 import { makeDeleteOneGroup } from './delete-one-group';
