+++ src/functions/macros/get-macros-by-group.spec.ts
@@ -7,13 +7,13 @@
  **************************************************************************/
 
 import { random } from 'lodash';
-import { CreatableGroup, CreatableMacro, isMacro, User } from '../../models';
+import { CreatableGroup, CreatableMacro, CreatableUser, isMacro, User } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { makeLoginOneUser } from '../auth/login-one-user';
 import { makeAddOneUserToManyGroups } from '../groups/add-one-user-to-many-groups';
 import { makeCreateOneGroup } from '../groups/create-one-group';
-import { CreatableUser, makeCreateOneUser, makeGetMyUser, makeGetOneUser } from '../users';
+import { makeCreateOneUser, makeGetMyUser, makeGetOneUser } from '../users';
 import { makeCreateOneMacro } from './create-one-macro';
 import { makeDeleteOneMacro } from './delete-one-macro';
 import { makeGetAllMacros } from './get-all-macros';
