+++ src/functions/scheduled-tasks/get-many-scheduled-scripts.spec.ts
@@ -7,11 +7,11 @@
  **************************************************************************/
 
 import { random } from 'lodash';
-import { isScheduledScript, User } from '../../models';
+import { CreatableUser, isScheduledScript, User } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { makeLoginOneUser } from '../auth/login-one-user';
-import { CreatableUser, makeCreateOneUser, makeGetOneUser } from '../users';
+import { makeCreateOneUser, makeGetOneUser } from '../users';
 import { makeCreateManyScheduledScripts } from './create-many-scheduled-scripts';
 import { makeDeleteAllScheduledScripts } from './delete-all-scheduled-scripts';
 import { makeGetAllScheduledScripts } from './get-all-scheduled-scripts';
