+++ src/models/user/creatable-user.ts
@@ -6,10 +6,12 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-export * from './creatable-user';
-export * from './is-valid-user';
-export * from './raw-creatable-user';
-export * from './raw-user';
-export * from './to-raw-creatable-user';
-export * from './to-user';
-export * from './user';
+import { UserRole } from './user';
+
+export interface CreatableUser {
+	user: string;
+	password: string;
+	name: string;
+	email: string;
+	role: UserRole;
+}
